sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "helm-sdk"
version = "0.3.0"
description = "Python SDK for Helm Protocol — AI Agent Gateway"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
]

[project.optional-dependencies]
identity = ["pynacl>=1.5"]
all = ["pynacl>=1.5"]

[project.urls]
Homepage = "https://github.com/anthropics/helm-gateway"
Documentation = "https://helm-protocol.org/docs"
